+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.1.0] 
 
 - Fix [#62](https://github.com/google/arb-editor/issues/62).
+- Fix quick fix of placeholder/key [#66](https://github.com/google/arb-editor/issues/66).
 
 ## [0.0.12]